Silence -Wconversion warnings with explicit narrowing casts
Twenty-two -Wconversion warnings (plus one -Wunused-parameter)
were emitting on every build. CMakeLists has -Wconversion enabled
deliberately; the noise was masking any new conversion bugs that
might appear later. Each site is an intentional narrowing where
the value is bounded by upstream logic; just make the narrowing
explicit so the compiler sees the intent.

Sites:
- packet_parser/utils.cpp: g_allocated_bytes (uint32_t) += size_t
- packet_parser/j2k_header.cpp: qcd->exponent[] (uint8_t) <- uint16_t>>11
- packet_parser/j2k_packet.cpp:
  - tag_tree_zero(): unused 'val' parameter -> [[maybe_unused]]
  - cblk->{zbp,lblock,npasses,bits_to_read} narrowing from int/int_log2()
    results, all bounded by checks earlier in the function
    (lblock+log2(passes) <= 16, npasses < JPEG2000_MAX_PASSES)
  - bnum (uint8_t) <- uint32_t b+1 (b bounded by num_bands <= 3)
- packet_parser/tile_handler.hpp: coc->NL (uint8_t) <- dfs.Idfs (uint32_t)
- frame_handler.hpp: start_SOD (uint32_t) += size (size_t)
- receiving_hook.cpp: stats arithmetic (size_t -> double) — hoist into
  a named const for readability

Both Release and Release+PARSER_OVERSHOOT_INSTR builds now compile
warning-free. Behavior unchanged.

+++ frame_handler.hpp
@@ -194,7 +194,7 @@ class frame_handler {
           tile_hndr.create(&cs);
           is_passed_header = 1;
         } else {
-          start_SOD += size;
+          start_SOD += static_cast<uint32_t>(size);
         }
         if (start_SOD == 0) {
           return;  // something wrong
